Update tutorial.md
grep "\t" perhaps intend to select only lines containing a tab. However, because string quote and escape does not work in standard shells like scripting languages and the result would be deleting lines not containing a t, which will result in difficult to understand errors loading the track on JBrowse.
A good gff file do not contain a line without tab other than comment lines starting with a #. So this part should be removed.